+++ CHANGELOG.md
@@ -6,3 +6,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - [ ] Use Tracing package for logging
+- [ ] Use [Nutype](https://github.com/greyblake/nutype) for Domain